02_Langchain_web-scrapping_Project/requirements.txt
streamlit
langchain
google-generativeai
faiss-cpu
python-dotenv
bs4
tqdm
